# contracts/contract3-api.openapi.yaml
openapi: 3.1.0
info:
  title: DarkPulse Investigator API
  version: 1.0.0
paths:
...
  /actors:
    get:
      tags:
      - Actors
      summary: List Actors
      operationId: list_actors_api_v1_actors_get
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Q
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /actors/{actor_id}:
    get:
      tags:
      - Actors
      summary: Get Actor
      operationId: get_actor_api_v1_actors__actor_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: actor_id
        in: path
        required: true
        schema:
          type: string
          title: Actor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiEnvelope:
      properties:
        data:
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/Pagination'
          - type: 'null'
        meta:
          additionalProperties: true
          type: object
          title: Meta
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
          title: Errors
      type: object
      title: ApiEnvelope
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
      type: object
      required:
      - code
      - message
      title: ErrorDetail
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Pagination:
      properties:
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        limit:
          type: integer
          title: Limit
          default: 50
        total:
          type: integer
          title: Total
          default: 0
      type: object
      title: Pagination
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8003